refactor: convert app to application factory pattern
